syncURL: example for HTTPS, OBEX channel selection
Use Google as example for HTTPS instead of Synthesis,
which is not a consumer-ready service.

Removed "-" in front of examples, lest some users think
that they are part of a valid syncURL.

Highlight the difference between the two OBEX examples
and explain about channel selection. The "+16" was
easy to miss before and not obvious.