D-Bus testing: add 'setupFiles' function in DBusUtil
The function 'setupFiles' is used to copy reference directory
trees from 'test/test-dbus' to 'xdg_root(./test-dbus)'. It
won't do copy unless own_xdg is enabled. Otherwise, user's
directories might be polluted.

This function is for those unit tests which are
hard(or impossbile) to do comparisons. To solve this
problem, we can create reference files in 'test/test-dbus'
and copy them to 'xdg_root' and compare expected results
with values returned from DBus server.

Currently 4 sub-directories are copied, but
maybe with different names, mappings are:
 test/test-dbus               ./test-dbus
    sync4j                       .sync4j
    data                         data
    config                       config
    cache                        cache