syncevo-dbus-server: fix an integer overflow in GetReports
The unit test 'testGetReportsByRef' in test-dbus.py
finds an integer overflow in GetReports. It will
call session.GetReports(2, 0xFFFFFFFF). It is possbile.
The sum of start and count is overflow. Thus
the 'if' statement is always false.
See the change:
if(index >= start && index < start + count) {
-->
if(index >= start && index - start < count) {