syncevo-dbus-server: send SYNC_DONE status after shutting down connection
This is based on an observation by Yongsheng. The previous code set
SYNC_DONE when the Synthesis engine signaled the corresponding event.
This is too early, SyncEvolution then still has work to do.

The second point where SYNC_DONE was set was after the return of
SyncContext::sync(). Closer, but then there's still the connection
which might have to be shut down.

In contrast to Yongsheng's patch, this one here also takes into account
that shutting down the connection might throw an exception and sets
SYNC_DONE both in that and the normal shutdown code path.