D-Bus API + syncevo-dbus-server: added Server.CheckSource()/GetDatabases() (MB #8091)
These operations do instantiate backends, but that should be possible
even if there is another active session using them. Because it is
useful for the sync-UI, let these two methods be called via the
read-only Server API.

I haven't added tests for them because that would create merge conflicts
with the "server-config" branch and shared configs. Should be added
later.