username/password: empty in templates
We used to have the strings "your SyncML server account name" and
"your SyncML server password" hard-coded in the code which generated
templates. As Jussi pointed out, this is a problem for GUIs which have
to hide these strings from their users.

The purpose of these strings was to have them in configs in cases
where the user forgot to set them, as a reminder.

This path moves the strings into the command line config setup,
thus achieving both of these goals.