D-Bus testing: setupFiles() + multiple snapshots
Different tests may need different files. setupFiles() now takes
a parameter that selects a subdirectory of $srcdir/test/test-dbus
for copying.

setupFiles() is meant to be used in tests with own_xdg=True. Using
it differently is an error, resulting in a test failure. Creating
the "test-dbus" directory is not necessary.