D-Bus API + testing + implementation: GetDatabases() requires source config
The implementation of GetDatabases() requires at least a temporary
config for the source. The previous implementation returned an empty
list when the source wasn't found, which cannot be distinguished from
the case that the source really has no databases.

Not reporting an error hid the misnaming of "todo" as "task" in
testGetDatabases(); name is fixed.

This patch changes the error reporting so that the same NoSuchSource
error is reported as in CheckSource().

This was found when adding "else" clauses to tests which expect
an exception to be thrown. Adding these clauses showed that some
tests succeeded despite not getting an error; all of these expected
the error incorrectly, so the try/except was removed.