SyncConfig::setConfigFilter(): per source-set properties were broken
The following code was meant to set m_peerNode and m_contextNode:
  m_peerNode->setFilter(filter);
  if (m_peerNode != m_contextNode) {
      m_peerNode->setFilter(filter);
  }

Setting m_peerNode twice was just a cut-and-paste error. The effect
of this was that shared sync properties were not read or written
at all.

Discovered when adding tests for shared "deviceId", test updates will
be committed together with that.