FilterConfigNode: existing properties not overwritten?!
According to the std::map::insert() documentation, it does not
overwrite existing values in the map. This is not what is intended
here, property values in the filter *should* supersede existing
values.

This was found when looking at the code and has not shown up
as a real failure yet?!