command line: preserve and show shared properties (MB #8048)
When adding a new peer, existing shared properties must be
preserved. The same applies to migrating (= copying) and old
configuration.

To be consistent, printing templates also needs to take existing
shared properties into account.

Special treatment is necessary for the "type" property: it must be
taken from the peer, not from the existing config, so that the peer
can choose a data format for himself. The peer config must be
consistent with regards to the backend for this to work.

The code is similar to syncevo-dbus-server, but not identical because
temporarily overriding properties is done differently. Therefore
SyncConfig::createServerTemplate() has been left as it is (= unaware
of shared properties).

The client-test SyncEvo::CmdlineTest was updated to cover some
of these case.