SyncContext: only use one engine instance, fixes slow sync issue
This removes code which was added as part of the some of the SAN
handling patches. The code added a second Synthesis engine while the
first one was still running, using this statement:

      // reinitializes the engine, only at this time can we decide whether
      // this is a server session or client session.
      SwapEngine swapengine(*this);

I recognize the comment. It must have been copied from an older
SyncEvolution revision. However, SyncEvolution no longer uses two
instances of the engine, so this code can and has to be removed again.
Instead the first instance is reinitialized once the full config is
known.

The effect was that both the old and the new instance had the file
with the anchor data open. Then the more up-to-date data was overwritten
by the older one without a valid anchor when it was deconstructed,
leading to a slow sync in each session.