DBus server: change the behavior of GetReports(MB#8049)
Reports returned to dbus clients are not what we
expect. Many keys are not included.

Current keys in dictionaries are conforming to
those in SyncML.cpp. They are defined formally in
BNF to be more accurate.

For statistics related reports, if value is 0,
it won't be put in the dictonaries to save time
and space.

Also refresh the unit test of GetReports in
test-dbus.py to make it successful.