shared config: when no context is given, search for config
A string like "scheduleworld" without context specified via @<context>
is meant to find such a config, regardless in which context it
is specified. The "default" context is meant to be searched first.

This patch adds this feature. It is done by a) determining whether a
context was specified explicitly and b) if not, searching the list of
configured peers.

Step a) requires a slightly modified normalizeConfigString()
implementation which tells the caller the context, even in the case
that @default gets stripped from the normalized form.

Step b) relies on sorting of the server list. Old-style configs or
configs in the default context come first and thus are found first, as
intended.

There's no explicit unit test for this. Running "client-test
Client::Sync" covers it (which is how the missing feature was
found...).