link problem: sysync::SySyncDebugPuts()
Because we are calling sysync::SySyncDebugPuts() directly again
instead of doing a dlsym() lookup (as in some earlier code, while
we had to load two different shared objects depending on client
or server mode), we also must link against libsynthesis in all
cases.

This could and should be reverted once MB #8264 gets resolved
and makes SySyncDebugPuts() a normal part of the libsynthesis
function table.