Fix a compiler warning
commit 607be1a4da4e4a385df56a69ef66db6e1de1d6f9 did not work as expected.
For the "strict aliasing" warning, it said:
"
Strict aliasing is an assumption, made by the C (or C++) compiler, that
dereferencing pointers to objects of different types will never refer to the
same memory location (i.e. alias each other.)
The exception to the rule is char*, which is allowed to point to any type.
"
So let's cast void* to char* instead of char* to void*.