Server alerted Sync: fix the timeout and signal handling during SAN
move initSAN from sync() to doSync()
It was once designed to put initSAN as early as possible to reduce
unnecessary resource if client does not respond (automatic sync).
However it lost the signal handling. Let's move it to doSync() at
the moment and come out the other solution for automatic sync later.

Timeout handler was not registerd for the transport, fixed.