autotools: install doc files in normal --docdir (MB #7168)
Probably because older automake didn't have a docdir, this was
only done for .deb files via checkinstall's doc_pak feature.
Now dist_doc_DATA is used for those files which may be of interest
to users of binary installations (.tar.gz, .deb, .rpm):
- README
- NEWS
- COPYING
- server readme's

The automatically generated ChangeLog is not installed because
it is considered too technical. It's also a lot larger than the
NEWS file.