Evolution Address Book: avoid picking CouchDB by default (MB #7877)
Append CouchDB address books at the end of the list,
otherwise preserving the order of address books.

The reason is Moblin Bugzilla #7877 (aka CouchDB
feature request #479110): the initial release of
evolution-couchdb in Ubuntu 9.10 is unusable because
it does not support the REV property.

Reordering the entries ensures that the CouchDB
address book is not used as the default database by
SyncEvolution, as it happened in Ubuntu 9.10.
Users can still pick it intentionally via
"evolutionsource".