autotools: fixed installation of backends with recent libtool
More recent libtool versions contain a check that the
install dir doesn't deviate too much from the libdir
compiled into the libs. This check failed when the
chosen path ended with a slash, leading to:

libtool: install: error: cannot install `syncecal.la' to a directory not ending in ...

I consider this a bug in libtool and sent a bug report
together with a fix. In the meantime, let's avoid the
problem by not adding any redundant slash in our
path...