defaultPeer: also shared when using legacy configs (MB #8334)
This patch allows a user to read and write the single "defaultPeer"
property in ~/.config/syncevolution/config.ini even when selecting
a peer which has a legacy config in ~/.config/syncevolution/<name>.
This is more convenient and ensures that the sync-UI works correctly
even in those cases where the legacy config cannot (or has not) been
migrated.

It works by adding multiplexing of properties to the HTTP_SERVER case
and then using a separate global config node.

Really, really old configs in .sync4j/evolution do not share the
defaultPeer property, but I think this is acceptable.