SyncContext::sync(): fixed potential incorrect memory access
"SyncReport buffer" was used as storage for later printing of the
report at the end of sync(). However, its scope ended earlier than
that. If the struct hadn't been entirely on the stack, we would have
had a problem.

With everything allocated on the stack, this wasn't a problem until I
added a std::string (next commit). Then it accessed a deconstructed
string => segfault.