SyncReport: record initial error (MB #7708)
The current use of recording the initial error ever printed during a
sync session is that it can be repeated at the end of the
"syncevolution" command line output or when printing sessions.

Initially the error is set to "synchronization process died prematurely".
Seeing that error in a session status.ini means that the report
wasn't updated because SyncContext::sync() never completed properly -
or is still running.

This patch also adds reseting of m_status in SyncReport::clear(),
which seemed to be missing.